% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_not.R
\name{group_by_not}
\alias{group_by_not}
\title{Group by all variables except those listed}
\usage{
group_by_not(.data, ...)
}
\arguments{
\item{.data}{a tbl}

\item{...}{Variables NOT to group by. Duplicated groups will be silently dropped.}
}
\value{
A grouped tbl
}
\description{
Opposite of dplyr::group_by()... for when it's more convenient to list the excluded variables
}
