% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcast.R
\name{ia_backcast}
\alias{ia_backcast}
\title{Use one data value series to backcast other series}
\usage{
ia_backcast(target, backcast_source, base_index = NULL)
}
\arguments{
\item{target}{Array of values to be backcast}

\item{backcast_source}{Array of value, whose growth rates will be applied to \code{target} as a backcast. \code{backcast_source[1]} aligns with \code{target[1]}. Must be same length as \code{target}}

\item{base_index}{Optional index to begin backcast in \code{target}}
}
\value{
An array of same length as \code{target} with backcast values.
}
\description{
Use one data value series to backcast other series
}
\examples{
ia_backcast(c(rep(NA, 5), 10:20), c(1:10))

ia_backcast(c(rep(NA, 5), 10:20), c(1:10), 10)
}
