% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_cast.R
\name{ia_cast}
\alias{ia_cast}
\title{Use one vector to fore- or backcast another vector}
\usage{
ia_cast(target, cast_source, cast_source_metric = "level",
  direction = "forward", base_index = NULL)
}
\arguments{
\item{target}{Vector of values to be fore- or backcast.}

\item{cast_source}{Vector of values or growth rates to be applied to \code{target} as a cast. \code{cast_source[1]} aligns with \code{target[1]}. Must be same length as \code{target}}

\item{cast_source_metric}{'level' to apply calculated from rates from \code{cast_source}. 'growth_rate' to use \code{cast_source[1]} as is.}

\item{direction}{'forward' to forecast from \code{base_index} or 'backward' to backcast from  \code{base_index}.}

\item{base_index}{Optional index to begin cast in \code{target}. Otherwise first/last non-NA value will be used}
}
\value{
An array of same length as \code{target} with backcast values.
}
\description{
Use one vector to fore- or backcast another vector
}
\examples{
ia_cast(c(10:20, rep(NA, 5)), c(1:16))

ia_cast(c(rep(NA, 5), 10:20), c(1:16), direction = "backward", base_index = 10)
}
