% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia_scale.R
\name{ia_scale}
\alias{ia_scale}
\title{A scale the sum or product of a vector to another value}
\usage{
ia_scale(input_array, target, method = "sum", reduce_target = FALSE)
}
\arguments{
\item{input_array}{Vector of values to be scaled.}

\item{target}{Desired result for the sum or product of the scaled result.}

\item{method}{'sum' to scale the sum of the totals, or 'product'.}

\item{reduce_target}{Logical. If \code{target} has more than one value, reduce it using the \code{method}}
}
\value{
An array of same length as \code{input_array}
}
\description{
A scale the sum or product of a vector to another value
}
\examples{
ia_scale(1:5, 20)

ia_scale(1:5, 20, method = "product")
}
