% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcast.R
\name{ip_backcast}
\alias{ip_backcast}
\title{Use one data value series to backcast other series}
\usage{
ip_backcast(datatable, series_target = "value",
  series_base = "value_base", growth_over = c("Year"))
}
\arguments{
\item{datatable}{A dataframe with two value columns}

\item{series_target}{The name of the series in \code{datatable} to be backcast.}

\item{series_base}{The name of the data series in \code{datatable} to use as backcast for missing \code{series_target}}

\item{growth_over}{Array of series names to calculate growth over. E.g. Year for annual percent growth}
}
\value{
A dataframe with the same dimensionality as \code{datatable}, with an additional data column of backcast data.
}
\description{
Use one data value series to backcast other series
}
