% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{ip_create_seed}
\alias{ip_create_seed}
\title{Create initial dataframe of all unique target series.}
\usage{
ip_create_seed(tars, names.exclude = c("value"), value.set = 1,
  value.name = "value")
}
\arguments{
\item{tars}{A list of data frames containing series to expand.}

\item{names.exclude}{Vector of names of series to exclude from the data frame creation.}

\item{value.set}{Single number or vector of numbers to initialize the seed \code{value.name}.}

\item{value.name}{Name of series in the data frame containing \code{value.set}.}
}
\value{
A data frame containing one row for each combination of the data frame series.
    The series in the first \code{tar} data frame vary fastest.
    The columns names match names supplied in each of the \code{tar} data frames.
}
\description{
Create initial dataframe of all unique target series.
Essentially \code{expand.grid()} for the tidyverse.
}
\examples{
tar1 <- data.frame(x = letters[1:2], value = runif(2))
tar2 <- data.frame(y = letters[3:5], value = runif(3))
ip_create_seed(list(tar1, tar2))
}
