% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ip_fit}
\alias{ip_fit}
\title{Iterative proportional fitting to scale seed values to multiple target subtotals.}
\usage{
ip_fit(datatable, targets, datatable.value.name = "value",
  target.value.names = "value", max.error = 0.01, max.iterations = 25,
  ice_cells = NULL, ice_cells.value.name = "value")
}
\arguments{
\item{datatable}{A data frame of values to be scaled to targets.}

\item{targets}{A list of data frames containing subtotal targets for the \code{datatable}. All totals in each target should be equal.}

\item{target.value.names}{The names of the series in \code{targets} containing subtotals to scale. Can be string or array of strings.}

\item{max.error}{The maximum total absolute difference allowed between final scaled values and targets.
Iterative scaling will complete once the error is below this threshold or \code{max.iterations} occur.}

\item{max.iterations}{The maximum number of iterations of scaling. Iterative scaling with end once this value is reached, even if the error is above \code{max.error}.}

\item{ice_cells}{Optional data frame of values with same series columns as \code{datatable}, specifying exact values to hit in the scaling.
Any values not listed, or \code{NA}s, will be scaled as normal.}

\item{ice_cells.value.name}{The name of the series of iced values in \code{ice_cells}.}

\item{datatable.value.names}{The name of the series in \code{datatable} containing the values to be scaled.}
}
\value{
A dataframe with the same dimensionality as \code{datatable}, with all values scaled to the subtotals specified in each data frame in \code{targets}.
}
\description{
Iterative proportional fitting to scale seed values to multiple target subtotals.
}
\examples{
tar1 <- data.frame(x = letters[1:2], value = c(50, 50))
tar2 <- data.frame(y = letters[3:5], value = c(20, 40, 40))
tar3 <- data.frame(z = letters[6:10], value = c(10, 20, 30, 40, 10))

tar.list <- list(tar1, tar2, tar3)
df <- ip_create_seed(tar.list) \%>\% ip_fit(tar.list)
}
