% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wrappers.R
\name{ip_fit_sl}
\alias{ip_fit_sl}
\title{Iterative proportional fitting to scale seed values to multiple target subtotals.}
\usage{
ip_fit_sl(datatable, targets, assumptions,
  datatable.value.name = "value", target.value.names = "value",
  assumption.value.names = c("value", "value_min", "value_max"),
  assumption.drop.names = c("Notes"), max.error = 0.01,
  max.iterations = 25, minmax.smash.param = 1/3, save.tars = FALSE,
  show.messages = TRUE)
}
\arguments{
\item{datatable}{A data frame of values to be scaled to targets.}

\item{targets}{A list of data frames containing subtotal targets for the \code{datatable}. All totals in each target should be equal.}

\item{assumptions}{A single data frame containing data belonging to \code{freeze_cells}, \code{freeze_slice}, \code{minmax_cells}, and \code{minmax_slice}.}

\item{target.value.names}{The names of the series in \code{targets} containing subtotals to scale. Can be string or array of strings.}

\item{assumption.value.names}{The names of the series in \code{assumptions} containing subtotals to scale. Three item array of the freeze value name, min value, max value.}

\item{assumption.drop.names}{The names of the series in \code{assumptions} to exclude from analysis. Usually columns containing notes.}

\item{max.error}{The maximum total absolute difference allowed between final scaled values and targets.
Iterative scaling will complete once the error is below this threshold or \code{max.iterations} occur.}

\item{max.iterations}{The maximum number of iterations of scaling. Iterative scaling with end once this value is reached, even if the error is above \code{max.error}.}

\item{minmax.smash.param}{Numeric value of  0 < x < 1. Following an out-of-bounds occurence for \code{minmax_cells}, the \code{minmax.smash.param} is the additional value added to the scaled value to bring it back into bounds.
Values close to 0 bind the value to the violated bound, while close to 1 bind the value to the other bound.
Values closer to 0 will require more iterations to complete.}

\item{datatable.value.names}{The name of the series in \code{datatable} containing the values to be scaled.}
}
\value{
A dataframe with the same dimensionality as \code{datatable}, with all values scaled to the subtotals specified in each data frame in \code{targets} and meeting criteria supplied in \code{assumptions}.
}
\description{
Iterative proportional fitting to scale seed values to multiple target subtotals.
}
