% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{ip_growth_transform}
\alias{ip_growth_transform}
\title{Scale data frame values to match growth targets over any series}
\usage{
ip_growth_transform(datatable, target, series_start = "value",
  series_target = "growth", series_growth_over = "growth__over",
  sep = " + ", growth_distance = 1)
}
\arguments{
\item{datatable}{A data frame of values.}

\item{target}{A data frame of target growth rates over one or more dimensions.
Should include an additional series to specify which series the growth are over.}

\item{series_start}{The name of the series in \code{datatable} to be compared to the \code{target}.}

\item{series_target}{The name of the \code{target} series containing growth rates.}

\item{series_growth_over}{The name of the series in \code{target} specifying the series that are covered by the growth rates.}

\item{sep}{Separator string to list multiple series in \code{series_growth_over}.}

\item{growth_distance}{Number of lags between observations when applying growth rates.}
}
\value{
A data frame of target values for use in \code{freeze_slice} in \code{ip_fit}.
}
\description{
Scale data frame values to match growth targets over any series
}
