% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_assumptions.R
\name{ip_load_assumptions}
\alias{ip_load_assumptions}
\title{Create all freeze and minmax assumption data frames from single data frame}
\usage{
ip_load_assumptions(assumptions, freeze.name = "value",
  minmax.name = c("value_min", "value_max"), drop.names = c("Notes"))
}
\arguments{
\item{assumptions}{A data frame of assumptions that will be converted into \code{freeze_cells}, \code{freeze_slice},
\code{minmax_cells}, and \code{minmax_slice} for \code{ip_expand()} and \code{ip_fit()} functions.}

\item{freeze.name}{The name of the series in \code{assumptions} with freeze values.}

\item{minmax.name}{The names of the series in \code{assumptions} with min and max values.}

\item{drop.names}{The name of the series in \code{assumptions} to remove from analysis.}
}
\value{
A list of data frames and list containing \code{freeze_cells}, \code{freeze_slice}, \code{minmax_cells}, and \code{minmax_slice}.
}
\description{
Create all freeze and minmax assumption data frames from single data frame
}
