% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{ip_load_mask_a}
\alias{ip_load_mask_a}
\title{Convert a data frame of seed masks to individual data frames}
\usage{
ip_load_mask_a(mask_raw, mask.value.name = "mask",
  drop.names = c("Notes"))
}
\arguments{
\item{mask_raw}{A data frame of seed masks.}

\item{mask.value.name}{The name of the series in \code{mask_raw} mask values.}

\item{drop.names}{The name of the series in \code{mask_raw} to remove from analysis.}
}
\value{
A list of data frames containing individual data frames for use in ip_mask_seed().
}
\description{
Convert a data frame of seed masks to individual data frames
Mostly for use in ip_mask_seed() function
}
