% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{ip_miss}
\alias{ip_miss}
\title{Compare data frame to target subtotals and calculate absolute error.}
\usage{
ip_miss(datatable, target, series_start = "value", series_target = "value")
}
\arguments{
\item{datatable}{A data frame of values previously scaled to targets.}

\item{target}{A data frame of target subtotals over one or more dimensions.}

\item{series_start}{The name of the series in \code{datatable} to be compared to the \code{target}.}

\item{series_target}{The name of the \code{target} series.}
}
\value{
A summarized data frame with the same dimensionality as \code{target}, with a measurement of error.
}
\description{
Compare data frame to target subtotals and calculate absolute error.
}
\examples{
df <- data.frame(x = rep(letters[1:2], 2), y = c(rep("c", 2), rep("d", 2)), value = runif(4))
tar1 <- data.frame(x = letters[1:2], value = c(20, 30))
df \%>\% ip_scale(tar1) \%>\% ip_miss(tar1)
}
