% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{ip_miss_a}
\alias{ip_miss_a}
\title{Compare data frame to target subtotals and calculate absolute error.}
\usage{
ip_miss_a(datatable, target_series, series_start = "value",
  series_target = "tar1")
}
\arguments{
\item{datatable}{A data frame of values previously scaled to targets.}

\item{target_series}{Subset of \code{names(datatable)} containing series with target subtotals.}

\item{series_start}{The name of the series in \code{datatable} to be compared to the \code{target}.}

\item{series_target}{The name of the \code{target} series.}
}
\value{
A summarized data frame with the same dimensionality as \code{target}, with a measurement of error.
}
\description{
Compare data frame to target subtotals and calculate absolute error in a single data frame.
Analogous to \code{ip_miss}.
Primarily for use within other \code{ipfitr} functions to increase speed.
}
