% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{ip_scale}
\alias{ip_scale}
\title{Single target scale function calling external dataframe.}
\usage{
ip_scale(datatable, target, series_start = "value", series_target = "value")
}
\arguments{
\item{datatable}{A dataframe of values to be scaled to target.}

\item{target}{A dataframe of target subtotals over one or more dimensions.}

\item{series_start}{The name of the series in \code{datatable} to be scaled to the \code{target}.}

\item{series_target}{The name of the \code{target} series.}
}
\value{
A dataframe with the same dimensionality as \code{datatable}, with \code{series_start} scaled to the subtotals specified by \code{series_target} .
}
\description{
Single target scale function calling external dataframe.
}
\examples{
df <- data.frame(x = rep(letters[1:2], 2), y = c(rep("c", 2), rep("d", 2)), value = runif(4))
tar1 <- data.frame(x = letters[1:2], value = c(20, 30))
ip_scale(df, tar1)

df \%>\% ip_scale(tar1, series_start = "count", series_target = "tar1")
}
