% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{ip_shares_calc}
\alias{ip_shares_calc}
\title{Scale data frame values to shares over specified series}
\usage{
ip_shares_calc(datatable, series_start = "value", groups = NULL)
}
\arguments{
\item{datatable}{A data frame of values.}

\item{series_start}{The name of the series in \code{datatable} to be converted to shares.}

\item{groups}{A character vector of series names to group data, excluding from share calculations.}
}
\value{
A summarized data frame with the same dimensionality as \code{datatable}, with values grouped by \code{groups}.
}
\description{
Scale data frame values to shares over specified series
}
