% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{ip_shares_tot}
\alias{ip_shares_tot}
\title{Create a generic share target of 1, dropping component series}
\usage{
ip_shares_tot(datatable, series_start = "value", groups = NULL)
}
\arguments{
\item{datatable}{A data frame of values.}

\item{series_start}{The name of the series in \code{datatable} to be converted to shares.}

\item{groups}{A character vector of series names to group data, will be retained in output.}
}
\value{
A data frame with reduced dimensionality from \code{datatable}, with values set to 1.
}
\description{
Create a generic share target of 1, dropping component series
}
