% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{ip_shares_transform}
\alias{ip_shares_transform}
\title{Scale data frame values to match share targets over any series}
\usage{
ip_shares_transform(datatable, target, series_start = "value",
  series_target = "share", series_share_of = "share__of",
  sep = " + ")
}
\arguments{
\item{datatable}{A data frame of values.}

\item{target}{A data frame of target shares over one or more dimensions.
Should include an additional series to specify which series the shares are over.}

\item{series_start}{The name of the series in \code{datatable} to be compared to the \code{target}.}

\item{series_target}{The name of the \code{target} series containing shares.}

\item{series_share_of}{The name of the series in \code{target} specifying the series that are covered by the shares.}

\item{sep}{Separator string to list multiple series in \code{series_share_of}.}
}
\value{
A data frame of target values for use in \code{freeze_slice} in \code{ip_fit}.
}
\description{
Scale data frame values to match share targets over any series
}
