% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{ip_mask_seed}
\alias{ip_mask_seed}
\title{Apply a layer of multiplicative masks to a seed.}
\usage{
ip_mask_seed(seed, mask, mask.value.name = "mask",
  drop.names = c("Notes"))
}
\arguments{
\item{seed}{A seed data frame, either supplied by the user or created using \code{ip_create_seed()}.}

\item{mask}{A data frame of seed masks.}

\item{mask.value.name}{The name of the series in \code{mask} with mask values.}

\item{drop.names}{The name of the series in \code{mask} to remove from analysis.}
}
\value{
A list of data frames containing individual data frames for use in ip_mask_seed().
}
\description{
Apply a layer of multiplicative masks to a seed. This will apply the mask to seed on all rows that match supplied elements.
Without changing the original seed, the mask is used to proportionally raise, lower, or zero seed values. This is an easy mechanism for flavoring the seed with research and manual edits.
The mask is an alternative to line-item changes using \code{mutate()} + \code{if_else()}.
}
